Update OperationScan to OperatorScan
Migrate to use updated conventions of Operator* classes that implement rx.Observable.Operator
